+++ app/src/js/components/Footer/footer.js
@@ -1,14 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import Popover from '../Popover/popover';
 const pckg = require('../../../../../package.json');
 
 const Footer = ({
   api,
   apiVersion,
+  cmrInfo,
 }) => {
   const DASH_VERSION = pckg.version;
   const { authenticated } = api;
   const { warning, versionNumber } = apiVersion;
+  const { cmrProvider, cmrEnv, cmrOauthProvider } = cmrInfo;
 
   const leftSideLinks = [
     { text: 'FOIA', url: 'https://www.nasa.gov/FOIA/index.html' },
@@ -37,7 +40,32 @@ const Footer = ({
             <li className="footer__version">
               <div className="version__label">Version</div>
               <div className="dashboard__version">Dashboard v{DASH_VERSION}</div>
-              <div className="api__version">API v{versionNumber}</div>
+              <div className="api__version">
+                <Popover
+                  className="popover--blue"
+                  id={'apiInformation'}
+                  placement="top"
+                  popover={true}
+                  target={`API v${versionNumber}`}
+                  popoverContent={
+                    <>
+                      <div className="popover-body--header">Cumulus API Configurations</div>
+                      <div className="popover-body--description">
+                        Below are details about your API Configurations that are associated to your
+                         DAAC and individual operator settings.<br/>
+                        <br/>
+                      </div>
+                      <div className="popover-body--main">{
+                        <ul className="popover-body--api-info">
+                          <li>CMR Environment: {cmrEnv}</li>
+                          <li>CMR Provider: {cmrProvider}</li>
+                          <li>CMR Authentication: {cmrOauthProvider || 'None'}</li>
+                        </ul>}
+                      </div>
+                    </>
+                  }
+                />
+              </div>
             </li>
             <li className="footer__opensource">
               {rightSideLinks.map((linkInfo, id) => (
@@ -54,7 +82,8 @@ const Footer = ({
 
 Footer.propTypes = {
   api: PropTypes.object,
-  apiVersion: PropTypes.object
+  apiVersion: PropTypes.object,
+  cmrInfo: PropTypes.object
 };
 
 export default Footer;