+++ app/src/js/utils/table-config/execution-status.js
@@ -1,3 +1,4 @@
+/* eslint-disable react/prop-types */
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import get from 'lodash/get';