+++ cypress/integration/granules_spec.js
@@ -49,7 +49,7 @@ describe('Dashboard Granules Page', () => {
 
       cy.contains('.table__filters--filter label', 'Recovery').find('span').click();
       cy.contains('.table__filters--filter label', 'Status').find('span').click();
-      cy.contains('.table__filters--filter label', 'Execution').find('span').click();
+      cy.contains('.table__filters--filter label', 'Duration').find('span').click();
       cy.contains('.table__filters--filter label', 'Name').find('span').click();
       cy.contains('.table__filters--filter label', 'Published').find('span').click();
 
@@ -57,7 +57,7 @@ describe('Dashboard Granules Page', () => {
 
       cy.get('.table .thead .tr .th').should('contain.text', 'Recovery');
       cy.get('.table .thead .tr .th').should('not.contain.text', 'Status');
-      cy.get('.table .thead .tr .th').should('not.contain.text', 'Execution');
+      cy.get('.table .thead .tr .th').should('not.contain.text', 'Duration');
       cy.get('.table .thead .tr .th').should('not.contain.text', 'Name');
       cy.get('.table .thead .tr .th').should('not.contain.text', 'Published');
 
@@ -159,15 +159,15 @@ describe('Dashboard Granules Page', () => {
               .should('have.attr', 'href')
               .and('be.eq', collectionHrefFromId(granule.collectionId));
 
-            // has link to provider
+            // Executions-list column has link to the executions list page
             cy.get('@columns').eq(5).children('a')
               .should('have.attr', 'href')
-              .and('be.eq', `/providers/provider/${granule.provider}`);
+              .and('be.eq', `/executions/executions-list/${encodeURIComponent(granule.collectionId)}/${encodeURIComponent(granule.granuleId)}`);
 
-            // Execution column has link to the detailed execution page
+            // has link to provider
             cy.get('@columns').eq(6).children('a')
               .should('have.attr', 'href')
-              .and('be.eq', `/executions/execution/${granule.execution.split('/').pop()}`);
+              .and('be.eq', `/providers/provider/${granule.provider}`);
 
             // Duration column
             cy.get('@columns').eq(7).invoke('text')
@@ -627,7 +627,7 @@ describe('Dashboard Granules Page', () => {
       cy.contains('button', 'Delete').click();
       cy.get('.default-modal.batch-async-modal ').as('modal');
 
-      cy.get('@modal').contains('div', 'Selection contains granules that are published to CMR which must be removed before deleting' && 'Remove published granules from CMR and delete?');
+      cy.get('@modal').contains('div', 'In order to delete these granules from Cumulus, we will first remove them from CMR' && 'Would you like to continue with your request');
     });
 
     it('Should show the correct DELETE modal when published granules are selected on the "processing granules" page', () => {
@@ -641,7 +641,7 @@ describe('Dashboard Granules Page', () => {
       cy.contains('button', 'Delete').click();
       cy.get('.default-modal.batch-async-modal ').as('modal');
 
-      cy.get('@modal').contains('div', 'Selection contains granules that are published to CMR which must be removed before deleting' && 'Remove published granules from CMR and delete?');
+      cy.get('@modal').contains('div', 'In order to delete these granules from Cumulus, we will first remove them from CMR' && 'Would you like to continue with your request');
     });
 
     it('Should show the correct DELETE modal when published granules are NOT selected', () => {
@@ -761,6 +761,14 @@ describe('Dashboard Granules Page', () => {
       cy.get('.scrollButtonLeft').trigger('mouseup', { button: 0 });
       cy.get('.scrollButtonLeft').trigger('mouseout');
       cy.get('.scrollButtonLeft').should('not.be.visible');
+    });
+
+    it('should navigate to executions-list when clicking on the executions-list link for a granule', () => {
+      cy.visit('/granules');
+      cy.url().should('include', 'granules');
+
+      cy.get('.tr:nth-child(1) > .td:nth-child(6) > a').click();
+      cy.url().should('include', 'executions-list');
     });
   });
 });